+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Use `passenger-ruby27` rather than `passenger-ruby26` as base image.
 - Copy the right `rails-console.sh` helper script into the image.
 - Fix setting the `PATH` variable to include the path to the `bin` directory.
+- The `rails` command in the `rails-console.sh` helper script was missing the
+  actual command.
 
 ## Version 2.0.1 (2020-09-13)